+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+* [home-manager/Services/Polybar] Refresh network data every second.
+
 ## [0.0.20] - 2024-01-01
 
 ### Added
@@ -1094,6 +1100,7 @@ You have to update it to:
 
 * Extraction from my personal configuration framework.
 
+[Unreleased]: https://github.com/ejpcmac/confkit/compare/main...develop
 [0.0.20]: https://github.com/ejpcmac/confkit/compare/v0.0.19...v0.0.20
 [0.0.19]: https://github.com/ejpcmac/confkit/compare/v0.0.18...v0.0.19
 [0.0.18]: https://github.com/ejpcmac/confkit/compare/v0.0.17...v0.0.18